Create a new test module for confound workflows.
