+++ tests/postprocess2/test_workflows.py
@@ -292,23 +292,6 @@ def test_resample_wf(
         helpers.plot_4D_img_slice(resampled_path, "resample.png")
 
 
-def test_build_get_scrub_targets_workflow(
-    artifact_dir, sample_confounds_timeseries, request, helpers
-):
-    """Test getting scrub targets from confounds given an FD threshold."""
-
-    test_path = helpers.create_test_dir(artifact_dir, request.node.name)
-    scrubbed_path = test_path / "scrubbed_confounds.tsv"
-
-    wf = build_scrubbing_workflow(
-        in_file=sample_confounds_timeseries,
-        out_file=scrubbed_path,
-        base_dir=test_path,
-        crashdump_dir=test_path,
-    )
-    wf.run()
-
-
 def test_scrubbing_wf(
     artifact_dir, sample_raw_image, plot_img, write_graph, request, helpers
 ):
